Refactoring of storage:
- main class Storage now is a "route" manager which main purpose is to
forward all requests to proper physical value manager
(PVMountPointManagerAdapter). It seems that lock-free tree could be
used.

- PVMountPointManager is a class that know everything about physical
volume layout and can correctly work with it. Also it can load and save
physical representation of volume.

- PVMountPointManagerAdapter is an adapter class to PVMountPointManager.
That means that it can adjust path if non root catalog of physical volume has
been mounted.

- PVManagerFactory only can create PVMountPointManagerAdapters. It
returns the proper adaptor for path in respect that every catalog of physical volumes can
be mounted.